requirements.txt
requests
terminaltables
python-dotenv